@@ -1,13 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.haberinadresi.androidapp"
-        minSdkVersion 19
-        targetSdkVersion 28
-        versionCode 24
-        versionName "2.4"
+        minSdkVersion 21
+        targetSdkVersion 29
+        versionCode 26
+        versionName "2.6"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables.useSupportLibrary = true
         multiDexEnabled true
