@@ -1,8 +1,7 @@
 package com.haberinadresi.androidapp.activities;
 
 import androidx.core.app.JobIntentService;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
+import androidx.lifecycle.ViewModelProvider;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -47,7 +46,7 @@ public class ColumnistSelectActivity extends AppCompatActivity implements Search
                 case "small":
                     setTheme(R.style.FontStyle_Small);
                     break;
-                case "medium":
+                case "medium": default:
                     setTheme(R.style.FontStyle_Medium);
                     break;
                 case "large":
@@ -56,8 +55,6 @@ public class ColumnistSelectActivity extends AppCompatActivity implements Search
                 case "xlarge":
                     setTheme(R.style.FontStyle_XLarge);
                     break;
-                default:
-                    setTheme(R.style.FontStyle_Medium);
             }
         } else {
             setTheme(R.style.FontStyle_Medium);
@@ -78,26 +75,22 @@ public class ColumnistSelectActivity extends AppCompatActivity implements Search
         recyclerView.setLayoutManager(new LinearLayoutManager(this, RecyclerView.VERTICAL, false));
         recyclerView.setHasFixedSize(true);
 
-        columnistAdapter = new ColumnistAdapter(this, new ArrayList<Columnist>());
+        columnistAdapter = new ColumnistAdapter(this, new ArrayList<>());
         recyclerView.setAdapter(columnistAdapter);
 
-        columnistSelectVM = ViewModelProviders.of(this).get(ColumnistSelectVM.class);
-        columnistSelectVM.getColumnistLiveData().observe(this, new Observer<List<Columnist>>() {
+        columnistSelectVM = new ViewModelProvider(this).get(ColumnistSelectVM.class);
+        columnistSelectVM.getColumnistLiveData().observe(this, columnistList -> {
 
-            @Override
-            public void onChanged(@Nullable List<Columnist> columnistList) {
+            columnistAdapter.setColumnistList(columnistList);
+            progressBar.setVisibility(View.INVISIBLE);
 
-                columnistAdapter.setColumnistList(columnistList);
-                progressBar.setVisibility(View.INVISIBLE);
-
-                // Start the service that deletes the outdated columnists from preferences
-                Intent updateMyColumnists = new Intent(ColumnistSelectActivity.this, UpdateMyColumnists.class);
-                Gson gson = new Gson();
-                String json = gson.toJson(columnistList);
-                updateMyColumnists.putExtra(getResources().getString(R.string.columnists_list), json);
-                JobIntentService.enqueueWork(ColumnistSelectActivity.this, UpdateMyColumnists.class,
-                        201, updateMyColumnists);
-            }
+            // Start the service that deletes the outdated columnists from preferences
+            Intent updateMyColumnists = new Intent(ColumnistSelectActivity.this, UpdateMyColumnists.class);
+            Gson gson = new Gson();
+            String json = gson.toJson(columnistList);
+            updateMyColumnists.putExtra(getResources().getString(R.string.columnists_list), json);
+            JobIntentService.enqueueWork(ColumnistSelectActivity.this, UpdateMyColumnists.class,
+                    201, updateMyColumnists);
         });
 
     }
